Support for driver ped animation (but bone rotations are not quite working)
If the capture includes the driver, the model name of the ped will be listed on the info panel, and a pointer property will be shown for selecting the armature.
The animation will be imported onto newly created empties that will be located in the same collection as the vehicle. The selected armature will attach to the animated empties via bone constraints.
Unfortunately, something appears to be wrong with the bone rotations.
StringProperties for the vehicle dummies were changed to PointerProperties.